+++ lib/src/main/kotlin/xyz/daimones/ktor/panel/Dataclasses.kt
@@ -47,6 +47,10 @@ package xyz.daimones.ktor.panel
  *
  * @property customLogoutTemplate Optional custom Mustache template for the logout confirmation page in the admin panel.
  *
+ * @property favicon Optional favicon URL for the admin panel. This allows you to set a custom icon.
+ *
+ * @property templateRenderer The template renderer/engine used to render templates.
+ *
  * @see Admin
  * @see BaseView
  */
@@ -63,6 +67,7 @@ data class Configuration(
     val customDetailsTemplate: String? = null,
     val customLoginTemplate: String? = null,
     val customDeleteTemplate: String? = null,
-    val customLogoutTemplate: String? = null
+    val customLogoutTemplate: String? = null,
+    val favicon: String = "/static/logo.ico",
+    val templateRenderer: TemplateRenderer = MustacheTemplateRenderer()
 )
-